Improved the efficiency of adding a message to an archive a bit.
Basically, we used to do copy some memory around to mine an MD5 hash without
having to change the archive's string pointer itself, to minimize the time we
locked the struct. But I realized sinc we writelock the archive before adding
a message anyway, no threads will even be accessing the structure while we mine
the hash, so we might as well save the time and the memory fragmentation.